Huff N Puff, Bite, Enemy timers
- It is no longer possible to create multiple Huff N Puffs. We wait
until the last HnP is gone before enabling the ability again.
- I’ve increased the force of HnP by 100f to make it a bit faster off
the mark. I’ve also made it a public variable rather than a hardcoded
value of 400f so we can tweak it more easily in future.
- Bite now has a cool down that we can control by a public variable
- Bite now returns a fraction of the enemy we are attacking’s total
health to BBW. The fraction is controlled by a variable called
lifestealDivider
- If BBW’s health is above it’s starting value, it will decrease until
it hits normal levels based on the healthDecay variable
- Fixed the cooldown/timer on the Troll and Woodsman attacks as they
were crazy quick