Improved synergy between CameraPanUp and CameraMove
Found some more elements in CameraPanUp.cs that didn’t agree with
CameraMove.cs. Mainly that CameraMove.cs locks the background to the
camera, but wasn’t active during CameraPanUp.cs’s load in movement due
to a bad interaction. What this meant was that a number of our scenes
had a background that was not initially aligned with the X axis of the
camera during load in, resulting in a weird snap when CameraMove.cs
took over.

We now fix the background to the x axis of the camera in CameraPanUp
too, so this no longer occurs.