Added platform movement switch
Added a switch to the MovingPlatformTestLevel scene that turns the
moving platform’s movement on and off.

Currently the switch is activated when LRRH runs over it.

It has a timeout after being hit that stops it being activated again
for a period. Additionally, it will only trigger again once LRRH moves
out of the trigger zone.